refactor winPercentage to return integer

+++ server/src/main/java/gui/matchday/dto/reports/ReportsResponseDTO.java
@@ -15,7 +15,7 @@ public class ReportsResponseDTO {
 
     private Integer matchesQuantity;
     private Integer winsQuantity;
-    private Double winPercentage;
+    private Integer winPercentage;
     private Team mostWatchedTeam;
     private Integer daysWithoutWatching;
